+++ .github/workflows/periodic.yml
@@ -11,6 +11,7 @@ jobs:
       matrix:
         os: [ubuntu-latest, macOS-latest, windows-latest]
         toolchain: [stable, beta, nightly]
+        features: ["", "compress"]
     runs-on: ${{ matrix.os }}
     steps:
       - uses: actions/checkout@v1
@@ -22,3 +23,4 @@ jobs:
       - uses: actions-rs/cargo@v1
         with:
           command: test
+          args: --features "${{ matrix.features }}"
