+++ src/main.rs
@@ -1,42 +1,48 @@
-use std::path::Path;
+use std::path::PathBuf;
 
-use clap::{crate_authors, crate_description, crate_name, crate_version, App, AppSettings, Arg};
+use clap::{AppSettings, Parser};
 
 use lindera_core::dictionary_builder::DictionaryBuilder;
-use lindera_unidic_builder::UnidicBuilder;
+use lindera_core::error::LinderaErrorKind;
+use lindera_core::LinderaResult;
+use lindera_unidic_builder::unidic_builder::UnidicBuilder;
 
-fn main() {
-    let app = App::new(crate_name!())
-        .setting(AppSettings::DeriveDisplayOrder)
-        .version(crate_version!())
-        .author(crate_authors!())
-        .about(crate_description!())
-        .help_message("Prints help information.")
-        .version_message("Prints version information.")
-        .version_short("v")
-        .arg(
-            Arg::with_name("INPUT_DIR")
-                .help("The directory where the UniDic source containing.")
-                .value_name("INPUT_DIR")
-                .required(true)
-                .takes_value(true),
-        )
-        .arg(
-            Arg::with_name("OUTPUT_DIR")
-                .help("The directory where the UniDic binary for Lindera is output.")
-                .value_name("OUTPUT_DIR")
-                .required(true)
-                .takes_value(true),
-        );
+/// Lindera UniDic dictionary builder
+#[derive(Parser, Debug)]
+#[clap(version, about, long_about = None, setting = AppSettings::DeriveDisplayOrder)]
+struct Args {
+    /// The dictionary source directory.
+    #[clap(short = 's', long = "dict-src", value_name = "DICT_SRC")]
+    dict_src: Option<PathBuf>,
 
-    let matches = app.get_matches();
+    /// The dictionary destination directory.
+    #[clap(short = 'd', long = "dict-dest", value_name = "DICT_DEST")]
+    dict_dest: Option<PathBuf>,
+}
 
-    let input_dir = Path::new(matches.value_of("INPUT_DIR").unwrap()).to_path_buf();
-    let output_dir = Path::new(matches.value_of("OUTPUT_DIR").unwrap()).to_path_buf();
+fn main() -> LinderaResult<()> {
+    env_logger::Builder::from_env(env_logger::Env::default().default_filter_or("info")).init();
 
-    let builder = UnidicBuilder::new();
-    match builder.build_dictionary(&input_dir, &output_dir) {
-        Ok(()) => println!("{}", "done"),
-        Err(msg) => println!("{}", msg),
+    let args = Args::parse();
+
+    let dict_builder = UnidicBuilder::new();
+
+    if args.dict_src.is_some() {
+        if args.dict_dest.is_some() {
+            let dict_src = args.dict_src.unwrap();
+            let dict_dest = args.dict_dest.unwrap();
+            match dict_builder.build_dictionary(&dict_src, &dict_dest) {
+                Ok(()) => (),
+                Err(msg) => {
+                    return Err(LinderaErrorKind::Args.with_error(anyhow::anyhow!(msg)));
+                }
+            }
+        } else {
+            return Err(LinderaErrorKind::Args.with_error(anyhow::anyhow!(
+                "`--dict-dest` is required when `--dict-src` is specified"
+            )));
+        }
     }
+
+    Ok(())
 }
